+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-08-24
+### Changed
+- All timestamps will be coerced into `Timestamp(MILLIS,true)` for better Flink compatibility
+### Fixed
+- Watermark computation bug
+
 ## [0.2.0] - 2023-07-24
 ### Changed
 - Upgraded to latest `datafusion` (fixes several known issues)